LayoutTests/fast/events/mutation-during-replace-child-expected.txt

Ensures that replaceChild() throws an exception if mutation even handler does something wrong

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.replaceChild(newChild, oldChild); threw exception Error: NotFoundError: DOM Exception 8.
PASS successfullyParsed is true

TEST COMPLETE
